requirements.txt
python-dotenv
Flask>=2.0
Flask-SQLAlchemy>=3.0
psycopg2-binary
sendgrid>=6.0
